New branch.

+++ libmaple/include/libmaple/libmaple_types.h
@@ -39,7 +39,7 @@ extern "C" {
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef long unsigned int uint32;
 typedef unsigned long long uint64;
 
 typedef signed char int8;
